tweak markdown formatting
